[26] Get server socket emitting to correct place in frontend.
